api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
supabase==2.1.1
pydantic==1.10.12 